Homework-7
Done! Teacher, I'm waiting next lesson!
